Create Role Repository

+++ Backend/DAA/DAA/src/main/java/com/nam/model/Role.java
@@ -15,10 +15,10 @@ import lombok.Setter;
 public class Role {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Enumerated(EnumType.STRING)
     @Column(length = 20)
     private ERole name;
-    
+
 }
